requirements.txt
Django==3.2
pyopendota==0.2.0
